rename pkg fundi to app
