+++ internal/app/doc.go
@@ -1,0 +1,2 @@
+// Package app contains routines in the application layer.
+package app
