--- internal/fundi/doc.go
@@ -1,2 +1,0 @@
-// Package fundi contains methods in the application layer.
-package fundi
